Add in study retrieval in control page
